server: keep static resources without host prefix
The Mongo database will not keep the host prefix, but instead
only the path to the resources will be kept. Only after a request
will be made the host prefix will be added.

For now the host prefix is http://apeserver.herokuapp.com/, but it
can be easily modified in the future by changin the host variable
from the server.js script. This way hosts can be easily changed without
the need to modifiy the database content.

Signed-off-by: Marius Avram <marius.avram1309@yahoo.ro>